### 2022-10-06 #4821 @editorialbot
[PRE REVIEW]: cosasi: Graph Diffusion Source Inference in Python
**Submitting author:** @lucasmccabe (<a href="http://orcid.org/0000-0002-7383-2823">Lucas McCabe</a>)
**Repository:** https://github.com/lmiconsulting/cosasi/
**Branch with paper.md** (empty if default branch): https://github.com/lmiconsulting/cosasi/tree/joss/
**Version:** 0.0.1
