### 2022-10-06 814371a
updated DOIs in bibfile
