+++ paper/references.bib
@@ -6,7 +6,8 @@
   number={8},
   pages={5163--5181},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tit.2011.2158885}
 }
 
 @article{zhu2016locating,
@@ -17,7 +18,8 @@
   number={5},
   pages={1217--1248},
   year={2016},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10618-015-0435-9}
 }
 
 @article{zhu2014information,
@@ -28,7 +30,8 @@
   number={1},
   pages={408--421},
   year={2014},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/ita.2013.6502991}
 }
 
 @inproceedings{prakash2012spotting,
@@ -37,7 +40,8 @@
   booktitle={2012 IEEE 12th International Conference on Data Mining},
   pages={11--20},
   year={2012},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/icdm.2012.136}
 }
 
 @article{ying2018diffusion,
@@ -48,7 +52,8 @@
   number={1},
   pages={1--95},
   year={2018},
-  publisher={Morgan \& Claypool Publishers}
+  publisher={Morgan \& Claypool Publishers},
+  doi={10.1007/978-3-031-79285-4}
 }
 
 @article{frkaszczak2022rpasdt,
@@ -58,7 +63,8 @@
   volume={17},
   pages={100988},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.softx.2022.100988}
 }
 
 @article{karczmarczyk2021oonis,
@@ -68,14 +74,16 @@
   volume={14},
   pages={100675},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.softx.2021.100675}
 }
 
 @article{miller2020eon,
   title={Eon (epidemics on networks): a fast, flexible python package for simulation, analytic approximation, and analysis of epidemics on networks},
   author={Miller, Joel C and Ting, Tony},
   journal={arXiv preprint arXiv:2001.02436},
-  year={2020}
+  year={2020},
+  doi={10.21105/joss.01731}
 }
 
 @software{lucas_mccabe_2021_4456181,
@@ -106,7 +114,8 @@
   number={1},
   pages={61--79},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.18637/jss.v084.i08}
 }
 
 @article{von2007tutorial,
@@ -117,5 +126,6 @@
   number={4},
   pages={395--416},
   year={2007},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11222-007-9033-z}
 }
