+++ README.md
@@ -1,5 +1,6 @@
 [![PyPI version](https://badge.fury.io/py/cosasi.svg)](https://badge.fury.io/py/cosasi)
 [![Documentation Status](https://readthedocs.org/projects/cosasi/badge/?version=latest)](https://cosasi.readthedocs.io/en/latest/?badge=latest)
+[![Downloads](https://pepy.tech/badge/cosasi)](https://pepy.tech/project/cosasi)
 [![MIT license](https://img.shields.io/badge/License-MIT-blue.svg)](https://lbesson.mit-license.org/)
 
 # cosasi (COntagion Simulation And Source Inference)
