### 2022-10-06 #4821 @lucasmccabe
Hello - per @editorialbot's request, I suggest the following potential reviewers from [the provided list](https://bit.ly/joss-reviewers):

- jonjoncardoso
- zoometh
- vc1492a
- leotrs
- akbaritabar
- ajgates42
- p-ortmann
- manmeetkaurbaxi
- sgeorge91
- gabett
- papachristoumarios
- jGaboardi
- dinacmistry
- sdmccabe  (FYSA: no relation)
- rodogi
- nick-gauthier

Thank you for your time and consideration!
