### 2022-10-10 7876c23
docs for 0.0.2 release
