### 2022-10-17 #4821 @danielskatz
@lucasmccabe - I want to let you know that JOSS has now switched to a track-based model of editors, and since this is in the track I'm leading (Computer Science, Information Science, and Mathematics). I'll be keeping an eye on it and looking for an available editor
