### 2022-10-18 #4821 @lucasmccabe
> I'm sorry @lucasmccabe, I'm afraid I can't do that. That's something only editors are allowed to do.

@danielskatz - Would it be possible to set the version to 0.0.2? A few updates were made since the initial submission. Thank you!
