### 2022-10-28 #4821 @danielskatz
👋 @jonjoncardoso, @zoometh, @vc1492a - would two of you be willing to review this submission for JOSS?
