### 2022-10-29 #4821 @sara-02
@danielskatz If possible, from mid-Nov, I can take this review. I have been working with the epidemic modeling library NDlib for some of my work, so I am familiar with the terminologies of this project.
