### 2022-10-31 #4894 @editorialbot
[REVIEW]: cosasi: Graph Diffusion Source Inference in Python
**Submitting author:** @lucasmccabe (<a href="http://orcid.org/0000-0002-7383-2823">Lucas McCabe</a>)
**Repository:** https://github.com/lmiconsulting/cosasi/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.4-joss
