### 2022-10-31 78e03b7
minor version + docs updates
Upgrade scipy and networx versions to address “AttributeError: module ‘scipy.sparse’ has no attribute ‘coo_array’” issue. Correct docstring in Earliest Infection First. Add code coverage visual.
