+++ requirements.txt
@@ -1,8 +1,8 @@
 ndlib==5.1.1
-networkx==2.8.4
+networkx==2.8.7
 numpy==1.21.4
 pytest==7.1.1
 scikit_learn==1.1.2
-scipy==1.9.0
+scipy==1.9.3
 coverage==6.3.2
 six==1.16.0
