+++ docs/developer_pages/releaselog.rst
@@ -3,12 +3,21 @@ Release Log
 =================
 
 
+0.0.3
+--------------------------------------
+Upgrade ``scipy`` and ``networx`` versions to address "module 'scipy.sparse' has no attribute 'coo_array'" issue_.
+Correct docstring in Earliest Infection First.
+Add code coverage visual.
+
 
 0.0.2
 --------------------------------------
-upgrade ``numpy`` version to address compatibility issue with ``threadpoolctl`` and ``scikit-learn`` exposed in spectral clustering
+Upgrade ``numpy`` version to address compatibility issue with ``threadpoolctl`` and ``scikit-learn`` exposed in spectral clustering.
 
 
 0.0.1
 --------------------------------------
-initial release
+Initial release
+
+
+.. _issue: https://github.com/pyg-team/pytorch_geometric/issues/4378
