+++ docs/conf.py
@@ -24,7 +24,7 @@ copyright = "2022, Lucas H. McCabe"
 author = "Lucas H. McCabe"
 
 # The full version, including alpha/beta/rc tags
-release = "0.0.2"
+release = "0.0.3"
 
 
 # -- General configuration ---------------------------------------------------
