### 2022-10-31 #4894 @lucasmccabe
Hello @danielskatz: I noticed that the version is listed as 0.0.1. Would it be possible to set this to 0.0.3? A few updates were made since the initial submission; these changes are minor, but are listed for transparency in our [release log](https://cosasi.readthedocs.io/en/latest/developer_pages/releaselog.html).

Also @sara-02 and @zoometh, if it helps at all regarding the code coverage statement, I included a [visual in the contributor guide](https://cosasi.readthedocs.io/en/latest/developer_pages/contributing.html#testing) for reference.

Thank you!
