### 2022-10-31 #4821 @jonjoncardoso
> 👋 @jonjoncardoso, @zoometh, @vc1492a - would two of you be willing to review this submission for JOSS?

Unfortunately, I won't have time this time
