### 2022-11-15 #4894 @danielskatz
👋 @zoometh - any update on your review?  It would be helpful if you can at least create your checklist via using the command `@editorialbot generate my checklist`. @editorialbot commands need to be the first thing in a new comment.
