### 2022-11-30 lmiconsulting/cosasi#1 @sara-02
Minor Documentation Update: JOSS review
One minor comment I have in the documentation is that the first quickstart examples present in readthedocs's main page https://cosasi.readthedocs.io/en/latest/index.html as well as the README here on GitHub has an image coded; it would be better to have the code in ```python``` format so that it can be copy pasted.
https://github.com/openjournals/joss-reviews/issues/4894
