### 2022-11-30 4edd38d
updated RTD documentation
