### 2022-11-30 800ac33
updated code snippet in README
