+++ README.md
@@ -17,14 +17,12 @@
 * [Table of Contents](#table-of-contents)
 * [Installation](#installation)
 * [Getting Started](#getting-started)
+* [Code Snippet](#code-snippet)
 * [Testing](#testing)
 * [Contributions](#contributions)
 * [Support](#support)
 * [Contact](#contact)
 * [License](#license)
-
-
-![carbon](./docs/_assets/carbon.png)
 
 
 ## Installation
@@ -44,6 +42,31 @@ pip install -r requirements.txt
 
 ## Getting Started
 Once `cosasi` is installed, feel free to review our [tutorial](https://cosasi.readthedocs.io/en/latest/tutorial.html) introducing major functionality. Official documentation, including a detailed [API reference](https://cosasi.readthedocs.io/en/latest/apiref.html), is available on [Read the Docs](https://cosasi.readthedocs.io/).
+
+
+## Code Snippet
+
+| ![carbon](./docs/_assets/carbon.png) |
+|:--:|
+| Above: [Carbon](https://github.com/carbon-app/carbon) image of example code snippet; copy-and-paste-able version below. |
+
+
+```python
+import networkx as nx
+import cosasi
+
+G = nx.fast_gnp_random_graph(100, 0.25)
+contagion = cosasi.StaticNetworkContagion(
+   G=G,
+   model="si",
+   infection_rate=0.01,
+   number_infected=3,
+)
+contagion.forward(100)
+I = contagion.get_infected_subgraph(step=15)
+result = cosasi.source_inference.multiple_source.netsleuth(G=G, I=I)
+result.evaluate(contagion.get_source())
+```
 
 ## Testing
 
