### 2022-11-30 lmiconsulting/cosasi#1 @lucasmccabe
Hi @sara-02 - great point. I have updated both the [README](https://github.com/lmiconsulting/cosasi#code-snippet) and the [narrative documentation](https://cosasi.readthedocs.io/en/latest/) to reflect your recommendation. I kept the [Carbon](https://github.com/carbon-app/carbon) image because I think it's aesthetically pleasing, but I added an equivalent code snippet in `python` format for easier copying.
