### 2022-11-30 #4894 @lucasmccabe
Thank you for your review @sara-02!  I have updated both the [README](https://github.com/lmiconsulting/cosasi#code-snippet) and the [narrative documentation](https://cosasi.readthedocs.io/en/latest/) to reflect your [recommendation](https://github.com/lmiconsulting/cosasi/issues/1).
