### 2022-12-01 lmiconsulting/cosasi#1 @lucasmccabe
Hi @sara-02 - are we comfortable closing this issue, or would you prefer an alternative corrective?
