### 2022-12-02 lmiconsulting/cosasi#1 @sara-02
> Hi @sara-02 - are we comfortable with my closing this issue, or would you prefer an alternative corrective?

Yes
