### 2022-12-07 #4894 @zoometh
👋 @lucasmccabe

* In the tutorial part

In general:
    - when using code with random generation (such as `nx.fast_gnp_random_graph(100, 0.25)`) it is useful to set up a pseudo-random number (e.g. `random.seed()`) so that the user can find the same results as the tutorial page.
     - This warning appears many times: `FutureWarning: laplacian_matrix will return a scipy.sparse array instead of a matrix in Networkx 3.0`

1. [Source Inference](https://cosasi.readthedocs.io/en/latest/tutorial_pages/source_inference.html#source-inference) page, the variables `I` and `G` are not instantiated. The following code snippets are not reproductible:

```
result = cosasi.single_source.netsleuth(I, G)
```
and
```
result = cosasi.multiple_source.netsleuth(I, G)
```

2. [When the Number of Sources is Unknown](https://cosasi.readthedocs.io/en/latest/tutorial_pages/estimating_source_number.html) page, this  code snippet creates warnings:

```
m = cosasi.utils.estimators.number_sources(I=I, number_sources=None, return_source_subgraphs=False, number_sources_method="eigengap")
```
gives:

```
<class 'networkx.utils.decorators.argmap'> compilation 12:4: FutureWarning: normalized_laplacian_matrix will return a scipy.sparse array instead of a matrix in Networkx 3.0.
/home/archesadmin/env/lib/python3.8/site-packages/cosasi/utils/estimators.py:25: FutureWarning: adjacency_matrix will return a scipy.sparse array instead of a matrix in Networkx 3.0.
  A = nx.adjacency_matrix(I)
/home/archesadmin/env/lib/python3.8/site-packages/sklearn/manifold/_spectral_embedding.py:259: UserWarning: Graph is not fully connected, spectral embedding may not work as expected.
  warnings.warn(
```

Same warnings for the following code snippets...

3. [Automatic Benchmarking](https://cosasi.readthedocs.io/en/latest/tutorial_pages/benchmarking.html?highlight=already_run_simulation#benchmark-from-simulation) page, the `already_run_simulation` is not instantiated
