### 2022-12-07 723333d
docs: more self-contained tutorial pages
