### 2022-12-07 cc99071
docs: updated source inference tutorial page to be more self-contained
