+++ cosasi/source_inference/tests/test_source_results.py
@@ -12,7 +12,22 @@ import numpy as np
 
 import cosasi
 
-from ..source_results import SourceResult, SingleSourceResult, MultiSourceResult
+from ..source_results import SourceResult, SingleSourceResult, MultiSourceResult, node_set_distance
+
+
+def test_node_set_distance():
+    G = nx.karate_club_graph()
+
+    # sets are individual nodes
+    assert node_set_distance(5, 2, G) == 2
+    # mixed
+    assert node_set_distance(5, [2, 3], G) == 6
+    assert node_set_distance([5, 6, 7], 2, G) == 7
+    # bother iterable
+    assert node_set_distance([5, 6], [2, 3], G) == 4
+    # overlapping
+    assert node_set_distance([5, 6], [5, 6], G) == 0
+    assert node_set_distance([5, 6], [2, 6], G) == node_set_distance(5, 2, G) == 2
 
 
 class Test_SourceResult(TestCase):
@@ -46,6 +61,15 @@ class Test_SourceResult(TestCase):
                 G="BAD INPUT",
             )
 
+    def test_bad_source_type_input(self):
+        with pytest.raises(ValueError):
+            assert SourceResult(
+                inference_method=self.inference_method,
+                scores=self.scores,
+                source_type="BAD INPUT",
+                G=self.G,
+            )
+
 
 class Test_SingleSourceResult(TestCase):
     def setUp(self):
@@ -63,6 +87,8 @@ class Test_SingleSourceResult(TestCase):
     def test_get_rank(self):
         assert self.result.get_rank(2) == 1
         assert self.result.get_rank(3) == 3
+        with pytest.raises(ValueError):
+            assert self.result.get_rank("BAD INPUT")
 
     def test_bad_graph_input(self):
         with pytest.raises(ValueError):
