### 2022-12-07 #4894 @lucasmccabe
Hi @zoometh, thank you for the careful review. You’ve raised several important points.

I think I've addressed your comments completely in my updates - please let me know if I've missed anything.

### Updates

1.	**Reproducibility of tutorial pages** (e.g. [Source Inference](https://cosasi.readthedocs.io/en/latest/tutorial_pages/source_inference.html), [When the Number of Sources is Unknown](https://cosasi.readthedocs.io/en/latest/tutorial_pages/estimating_source_number.html), [Automatic Benchmarking](https://cosasi.readthedocs.io/en/latest/tutorial_pages/benchmarking.html)) -

    This is now addressed in the most recent documentation update. I went through the tutorial and ensured that all variable references are self-contained in that page. I also added command line outputs where reasonable so that users may check that their results line up.

2.	**Setting a random seed**:

    This is addressed in two ways:

    - In the tutorial, random graph generation is now seeded. `NetworkX` uses both `random` and `np.random` for number generation, so I seed with both.

    - There is also stochasticity in the diffusion processes - both in infection initialization and spread. I updated the library to expose a `seed` parameter in the contagion simulator; this is reflected in package version [0.0.4](https://pypi.org/project/cosasi/) and throughout the documentation.

    In the subsequent section (Seeding Confirmation), I run a few checks to confirm that things are properly seeded in the updated version.

3.	**Warnings**:

    These are mostly things like future warnings coming directly from our dependencies (e.g. *FutureWarning: laplacian_matrix* is hard-coded [here](https://networkx.org/documentation/stable/_modules/networkx/linalg/laplacianmatrix.html)). They should not reflect risk to the user, since dependencies have been fixed in the requirements file.

    Nonetheless, these warnings will be a hassle to users, so I have addressed them in package version [0.0.4](https://pypi.org/project/cosasi/). Using this updated version, I went through the tutorial again and experienced no warnings, errors, etc.

### Seeding Confirmation

Imports, seeds, etc. used in this section:

```python
import networkx as nx, cosasi, random, numpy as np
seed = 42
random.seed(seed)
np.random.seed(seed)
```

Double-check seeding of random graph generation:

```python
>>> G = nx.fast_gnp_random_graph(200, 0.15, seed=seed)
>>> H = nx.fast_gnp_random_graph(200, 0.15, seed=seed)
>>> G.edges == H.edges
True
```

Double-check seeding of diffusion simulation:

```python
>>> contagion1 = cosasi.StaticNetworkContagion(G=G, model="si", infection_rate=0.01, fraction_infected=0.05, seed=seed)
>>> contagion1.forward(steps=10)
>>> contagion2 = cosasi.StaticNetworkContagion(G=G, model="si", infection_rate=0.01, fraction_infected=0.05, seed=seed)
>>> contagion2.forward(steps=10)
>>>
>>> contagion1.get_source() == contagion2.get_source()
True
>>> contagion1.get_infected_subgraph(step=5).edges == contagion2.get_infected_subgraph(step=5).edges
True
```
