### 2022-12-07 #4894 @lucasmccabe
Hello @danielskatz: Would it be possible to set the version to [0.0.4](https://pypi.org/project/cosasi/)? To address some of @zoometh's comments, I packaged an updated version.

These changes are minor, but are listed for transparency in our [release log](https://cosasi.readthedocs.io/en/latest/developer_pages/releaselog.html#id1).
