### 2022-12-10 #4894 @lucasmccabe
@zoometh Yes, I became aware of this paper recently, as well, and will update mine to cite it.

That said, SLVAE fills a different niche than cosasi - SLVAE is an implementation of a single autoencoder-based strategy, whereas cosasi is a more general framework.
