### 2022-12-10 c6e8542
SL-VAE reference
