+++ paper/references.bib
@@ -1,3 +1,12 @@
+@inproceedings{ling2022source,
+  title={Source Localization of Graph Diffusion via Variational Autoencoders for Graph Inverse Problems},
+  author={Ling, Chen and Jiang, Junji and Wang, Junxiang and Liang, Zhao},
+  booktitle={Proceedings of the 28th ACM SIGKDD Conference on Knowledge Discovery and Data Mining},
+  pages={1010--1020},
+  year={2022},
+  doi={10.1145/3534678.3539288}
+}
+
 @article{shah2011rumors,
   title={Rumors in a network: Who's the culprit?},
   author={Shah, Devavrat and Zaman, Tauhid},
