### 2022-12-10 #4894 @lucasmccabe
@zoometh The paper has been updated to cite SL-VAE. Thank you!
