+++ paper/references.bib
@@ -8,7 +8,7 @@
 }
 
 @article{shah2011rumors,
-  title={Rumors in a network: Who's the culprit?},
+  title={Rumors in a Network: Who’s the Culprit?},
   author={Shah, Devavrat and Zaman, Tauhid},
   journal={IEEE Transactions on information theory},
   volume={57},
@@ -20,7 +20,7 @@
 }
 
 @article{zhu2016locating,
-  title={Locating the contagion source in networks with partial timestamps},
+  title={Locating the Contagion Source in Networks With Partial Timestamps},
   author={Zhu, Kai and Chen, Zhen and Ying, Lei},
   journal={Data Mining and Knowledge Discovery},
   volume={30},
@@ -32,7 +32,7 @@
 }
 
 @article{zhu2014information,
-  title={Information source detection in the SIR model: A sample-path-based approach},
+  title={Information Source Detection in the SIR Model: A Sample-Path-Based Approach},
   author={Zhu, Kai and Ying, Lei},
   journal={IEEE/ACM Transactions on Networking},
   volume={24},
@@ -44,7 +44,7 @@
 }
 
 @inproceedings{prakash2012spotting,
-  title={Spotting culprits in epidemics: How many and which ones?},
+  title={Spotting Culprits in Epidemics: How Many and Which Ones?},
   author={Prakash, B Aditya and Vreeken, Jilles and Faloutsos, Christos},
   booktitle={2012 IEEE 12th International Conference on Data Mining},
   pages={11--20},
@@ -54,7 +54,7 @@
 }
 
 @article{ying2018diffusion,
-  title={Diffusion source localization in large networks},
+  title={Diffusion Source Localization in Large Networks},
   author={Ying, Lei and Zhu, Kai},
   journal={Synthesis Lectures on Communication Networks},
   volume={11},
@@ -66,7 +66,7 @@
 }
 
 @article{frkaszczak2022rpasdt,
-  title={RPaSDT—Rumour RPaSDT—Rumor Propagation and Source Detection Toolkit},
+  title={RPaSDT — Rumor Propagation and Source Detection Toolkit},
   author={Fr{\k{a}}szczak, Damian},
   journal={SoftwareX},
   volume={17},
@@ -77,7 +77,7 @@
 }
 
 @article{karczmarczyk2021oonis,
-  title={OONIS—Object-oriented network infection simulator},
+  title={OONIS — Onis—Object-Oriented Network Infection Simulator},
   author={Karczmarczyk, Artur and Jankowski, Jaros{\l}aw and W{\k{a}}tr{\'o}bski, Jaros{\l}aw},
   journal={SoftwareX},
   volume={14},
@@ -88,7 +88,7 @@
 }
 
 @article{miller2020eon,
-  title={Eon (epidemics on networks): a fast, flexible python package for simulation, analytic approximation, and analysis of epidemics on networks},
+  title={Eon (Epidemics on Networks): A Fast, Flexible Python Package for Simulation, Analytic Approximation, and Analysis of Epidemics on Networks},
   author={Miller, Joel C and Ting, Tony},
   journal={arXiv preprint arXiv:2001.02436},
   year={2020},
@@ -107,7 +107,7 @@
 }
 
 @article{jenness2018epimodel,
-  title={EpiModel: an R package for mathematical modeling of infectious disease over networks},
+  title={EpiModel: Epimodel: An R Package For Mathematical Modeling of Infectious Disease Over Networks},
   author={Jenness, Samuel M and Goodreau, Steven M and Morris, Martina},
   journal={Journal of statistical software},
   volume={84},
@@ -116,7 +116,7 @@
 }
 
 @article{rossetti2018ndlib,
[…]
