### 2022-12-11 #4894 @lucasmccabe
@danielskatz Thank you for providing these next steps. A release with tag `cosasi-0.0.4` is available [here](https://github.com/lmiconsulting/cosasi/releases/tag/cosasi-0.0.4).

As for Zenodo archiving, I will request that my organization enable third-party access to the Zenodo application. I will follow up when this is done.
