### 2022-12-13 #4894 @lucasmccabe
@danielskatz

- I made a new release for Zenodo - its tag is `v0.0.4-joss`, available [here](https://github.com/lmiconsulting/cosasi/releases/tag/v0.0.4-joss).
- The Zenodo archive has DOI [10.5281/zenodo.7430558](https://doi.org/10.5281/zenodo.7430558).
- I checked the metadata, which appears to list the title aligned with the paper title, correct author list, etc.

Please let me know if I've missed anything!

![image](https://user-images.githubusercontent.com/12187602/207205673-aa560d45-4239-4396-9550-18c43bc32b40.png)
