+++ paper/references.bib
@@ -110,7 +110,7 @@
 }
 
 @article{jenness2018epimodel,
-  title={EpiModel: An R Package For Mathematical Modeling of Infectious Disease Over Networks},
+  title={EpiModel: An {R} Package For Mathematical Modeling of Infectious Disease Over Networks},
   author={Jenness, Samuel M and Goodreau, Steven M and Morris, Martina},
   journal={Journal of Statistical Software},
   volume={84},
